requirements.txt
langchain
langchain-community
langchain-openai
openai
pandas
streamlit
python-dotenv
faiss-cpu